backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
joblib>=1.4.2
numpy>=1.26.4
scikit-learn>=1.5.0
xgboost>=2.1.0
pillow>=10.4.0
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.44.0
google-genai>=1.0.0
python-dotenv>=1.0.1
requests>=2.32.0
gtts>=2.5.0
